% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ald_density}
\alias{ald_density}
\title{Asymmetric-Laplace density at the tau-th quantile parametrisation
(Yu & Moyeed 2001), scale \code{sigma}. Vendored from the mixqr ALD idiom.}
\usage{
ald_density(u, sigma, tau)
}
\description{
Asymmetric-Laplace density at the tau-th quantile parametrisation
(Yu & Moyeed 2001), scale \code{sigma}. Vendored from the mixqr ALD idiom.
}
\keyword{internal}
