% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{align_labels}
\alias{align_labels}
\title{Align a replicate's regimes to the reference by matching constant coefficients.}
\usage{
align_labels(bc, ref)
}
\description{
Align a replicate's regimes to the reference by matching constant coefficients.
}
\keyword{internal}
