% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spmixqr.R
\name{build_car_slot}
\alias{build_car_slot}
\title{Assemble the obj$car slot from the fitted reduced CAR coefficients.}
\usage{
build_car_slot(beta, car_block, car_obj, car, car_alpha, lam_phi)
}
\description{
Assemble the obj$car slot from the fitted reduced CAR coefficients.
}
\keyword{internal}
