% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep-coef.R
\name{coef_objective_sp}
\alias{coef_objective_sp}
\title{Sparse-aware smoothed penalised objective (Matrix-friendly).}
\usage{
coef_objective_sp(b, Xt, y, tau, w, Pen, h)
}
\description{
Sparse-aware smoothed penalised objective (Matrix-friendly).
}
\keyword{internal}
