% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spmixqr.R
\name{compute_edf}
\alias{compute_edf}
\title{Effective degrees of freedom (component smoothers + gate smoother).}
\usage{
compute_edf(fit, Xt, Z, Pen_beta, Pen_gamma, G, des, spatial_coef)
}
\description{
Effective degrees of freedom (component smoothers + gate smoother).
}
\keyword{internal}
