% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{coords_subset}
\alias{coords_subset}
\title{Subset coordinates (matrix or region vector) by a logical index.}
\usage{
coords_subset(coords, idx)
}
\description{
Subset coordinates (matrix or region vector) by a logical index.
}
\keyword{internal}
