% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.dens_floor}
\alias{.dens_floor}
\title{Numeric floor for densities / probabilities (avoid log(0)).}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
.dens_floor
}
\description{
Numeric floor for densities / probabilities (avoid log(0)).
}
\keyword{internal}
