% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep-gate.R
\name{gate_predict}
\alias{gate_predict}
\title{Gate probabilities pi (n x G) from gate coefficients and design.}
\usage{
gate_predict(gamma, Z)
}
\description{
Gate probabilities pi (n x G) from gate coefficients and design.
}
\keyword{internal}
