% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{incidence_matrix}
\alias{incidence_matrix}
\title{Unit-incidence matrix R (n x L), sparse, from an observation->unit index.}
\usage{
incidence_matrix(unit_idx, L)
}
\arguments{
\item{unit_idx}{Length-n integer index of each observation's unit (1..L).}

\item{L}{Number of units.}
}
\value{
A sparse \verb{n x L} incidence matrix.
}
\description{
Unit-incidence matrix R (n x L), sparse, from an observation->unit index.
}
\keyword{internal}
