% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_starts}
\alias{init_starts}
\title{A list of \code{nstart} initial responsibility matrices cycling through start types.}
\usage{
init_starts(y, X, coords, G, tau, nstart)
}
\description{
A list of \code{nstart} initial responsibility matrices cycling through start types.
}
\keyword{internal}
