% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isTRUE_vec}
\alias{isTRUE_vec}
\title{Vectorised isTRUE: TRUE where x is TRUE, FALSE for FALSE/NA.}
\usage{
isTRUE_vec(x)
}
\description{
Vectorised isTRUE: TRUE where x is TRUE, FALSE for FALSE/NA.
}
\keyword{internal}
