% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nngp.R
\name{matern_cor}
\alias{matern_cor}
\title{Matern correlation (closed forms for nu = 0.5 or 1.5).}
\usage{
matern_cor(h, range, nu = 0.5)
}
\description{
Matern correlation (closed forms for nu = 0.5 or 1.5).
}
\keyword{internal}
