% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nngp.R
\name{maxmin_order}
\alias{maxmin_order}
\title{Max-min ordering of points (Guinness 2018): greedily pick the point farthest
(in min-distance) from those already chosen. O(n^2) pure R; fine to a few thousand.}
\usage{
maxmin_order(coords)
}
\description{
Max-min ordering of points (Guinness 2018): greedily pick the point farthest
(in min-distance) from those already chosen. O(n^2) pure R; fine to a few thousand.
}
\keyword{internal}
