% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{moran_unit_index}
\alias{moran_unit_index}
\title{Observation -> unit index for the Moran aggregation.}
\usage{
moran_unit_index(object, spqw)
}
\description{
Observation -> unit index for the Moran aggregation.
}
\keyword{internal}
