% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{nb_to_named_list}
\alias{nb_to_named_list}
\title{Convert an spdep nb/listw to mgcv's named-list neighbour form.}
\usage{
nb_to_named_list(areal, lev)
}
\description{
Convert an spdep nb/listw to mgcv's named-list neighbour form.
}
\keyword{internal}
