% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{perm_moran}
\alias{perm_moran}
\title{Permutation Moran's I for a unit-level vector against a weights matrix.}
\usage{
perm_moran(x, W, nsim = 999L)
}
\description{
Permutation Moran's I for a unit-level vector against a weights matrix.
}
\keyword{internal}
