% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_car_phi}
\alias{predict_car_phi}
\title{Broadcast the CAR phi surface (L x G) to new observations (n x G) via their unit.}
\usage{
predict_car_phi(object, newcoords)
}
\description{
Broadcast the CAR phi surface (L x G) to new observations (n x G) via their unit.
}
\keyword{internal}
