% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_designs}
\alias{predict_designs}
\title{Build component / gate designs at new data (same column layout as the fit).}
\usage{
predict_designs(object, newdata, newcoords)
}
\description{
Build component / gate designs at new data (same column layout as the fit).
}
\keyword{internal}
