% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densities.R
\name{psi_smooth}
\alias{psi_smooth}
\title{Smoothed check-loss score (gradient w.r.t. the residual).}
\usage{
psi_smooth(u, tau, h)
}
\description{
Smoothed check-loss score (gradient w.r.t. the residual).
}
\keyword{internal}
