% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{resample_blocks}
\alias{resample_blocks}
\title{Resample whole spatial blocks (with replacement) to length ~ n.}
\usage{
resample_blocks(blocks)
}
\description{
Resample whole spatial blocks (with replacement) to length ~ n.
}
\keyword{internal}
