% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rho_tau}
\alias{rho_tau}
\title{Check (pinball) loss: u times (tau minus the indicator that u is negative).}
\usage{
rho_tau(u, tau)
}
\description{
Check (pinball) loss: u times (tau minus the indicator that u is negative).
}
\keyword{internal}
