% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em-spatial.R
\name{spatial_em_fit}
\alias{spatial_em_fit}
\title{One penalised-EM fit from an initial responsibility matrix.}
\usage{
spatial_em_fit(
  y,
  Xt,
  Z,
  G,
  tau,
  method,
  p_init,
  Pen_beta,
  Pen_gamma,
  h,
  spatial_coef,
  kdectrl,
  control
)
}
\description{
One penalised-EM fit from an initial responsibility matrix.
}
\keyword{internal}
