% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spmixqr.R
\name{extract_const}
\alias{extract_const}
\title{Constant coefficient matrix (p x G) from the augmented coefficients.}
\usage{
extract_const(beta, des, p)
}
\description{
Constant coefficient matrix (p x G) from the augmented coefficients.
}
\keyword{internal}
