% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spmixqr.R
\name{resolve_coords}
\alias{resolve_coords}
\title{Resolve coordinates / areal specification.}
\usage{
resolve_coords(coords, areal, data, n)
}
\description{
Resolve coordinates / areal specification.
}
\keyword{internal}
