% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{symmetrise}
\alias{symmetrise}
\title{Symmetrise a matrix (numerical hygiene for covariances).}
\usage{
symmetrise(A)
}
\description{
Symmetrise a matrix (numerical hygiene for covariances).
}
\keyword{internal}
